geom/tri: Adding debug prints and corrections
At the end of the work here, I need to ask what to do with all the prints anyway.
For debugging they are really nice to have, but as far as I can see it is the intention to keep all the debug prints away from the code.
Furthermore there is no logging module if I'm not wrong. Therefore I think I'll wipe them out at the end.

Anyway, fixed the one of the sub-checks inside check no. 2. I assumed that the children's triangle is pointing into the same directions as the connected non-coplanar triangle of the "parent" triangle.
However, when inspecting the mesh again, I noticed that this is not the case, so I commonized the algorith in that way that it doesn't matter where the children's triangle is placed, but that the angle is the same.
That was the intended check to be honest and this is typical to the stacked cylinders.

Signed-off-by: Thomas Karl Pietrowski <thopiekar@googlemail.com>