geom/tri: Tweaking the _COPLANAR_ANGLE
Just tweaked the value to 0.004, since I noticed that the triangle_filter fails at the end of our triangle_filter.
According to the investigations made, the condition:

plane.vector_angle(triangle.normal) <= coplanar_angle

is not met. As a logical I consequence I removed the condition first to understand the technical need here and noticed it is meant to filter small triangles, which are around the holes corner.
Therefore, I increased the value in steps, tested with the models provided and didn't notice any malfunctions elsewhere.

Fixes an issue reported at: https://github.com/tetonsim/is-cura-ui/issues/133

Signed-off-by: Thomas Karl Pietrowski <thopiekar@googlemail.com>