geom/tri: Adding concave/convex grow algorithm
Tried to reuse as many functions and code pieces as possible.
The algorithm is looking for concave (or convex) as well as planar triangles in the neighborhood.
The planar ones are catched for the reason that holes are simplified during tesselation into squares.
However, since our messages are made using triangles, these square get split into two triangles.
For this reason we need to allow to grow in this direction to keep looking for following concave triangles.

Contributes to https://github.com/tetonsim/is-cura-ui/issues/133